+++ tests/__init__.py
@@ -1,0 +1,2 @@
+# tests/init.py
+from .tools import *
