+++ tests/tools/__init__.py
@@ -1,0 +1,5 @@
+# __init__.py
+from .single_test import single_benchmark_execution
+from .multi_test import mass_benchmark_execution
+from .local_server import start_server
+from .run_benchmark import original_run_benchmark
