+++ tests/tools/run_benchmark.py
@@ -1,0 +1,53 @@
+# run_benchmark.py
+import logging
+from typing import Dict
+from pathlib import Path
+import subprocess
+import time
+import logging
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+)
+logger = logging.getLogger(__name__)
+
+def original_run_benchmark(
+    params: Dict[str, int],
+    model: str,
+    port: int,
+    benchmark_script: str,
+    result_filename: Path,
+) -> None:
+    """Run a single benchmark with the given parameters."""
+    # fmt: off
+    cmd = [
+        "python", benchmark_script,
+        "--backend", "vllm",
+        "--model", model,
+        "--port", str(port),
+        "--dataset-name", "random",
+        "--num-prompts", str(params["num_prompts"]),
+        "--random-input-len", str(params["input_len"]),
+        "--random-output-len", str(params["output_len"]),
+        "--ignore-eos",  # Ignore EOS tokens to force max output length as set
+        "--percentile-metrics", "ttft,tpot,itl,e2el",  # must add e2el in order for it to be logged
+        "--save-result",
+        "--result-filename", str(result_filename)
+    ]
+    # fmt: on
+
+    logger.info(f"Running benchmark with parameters: {params}")
+    logger.info(f"Command: {' '.join(cmd)}")
+
+    try:
+        subprocess.run(cmd, check=True)
+        logger.info("Benchmark completed successfully")
+    except subprocess.CalledProcessError as e:
+        logger.error(f"Benchmark failed with error: {e}")
+    except Exception as e:
+        logger.error(f"Unexpected error during benchmark: {e}")
+
+    # Add a small delay between runs to ensure system stability
+    time.sleep(2)
+
